Fix condition where the same entry is made twice.
When making a rule in a security group, if the rule is the same, terraform will try and make it twice, but AWS will consolidate out the duplicate.  This fix does that before applying it to AWS so that whenever terraform is run it won't try to remove the rule with one entry and the replace it again with the same rule twice.